Move upsert/delete tenant functions to `tenant-rmw-sql`
