+++ packages/pn-consumers/test/utils.ts
@@ -5,10 +5,8 @@ import {
 } from "pagopa-interop-commons-test";
 import { afterEach, inject } from "vitest";
 import { Purpose, Tenant } from "pagopa-interop-models";
-import {
-  purposeReadModelServiceBuilder,
-  tenantReadModelServiceBuilder,
-} from "pagopa-interop-readmodel";
+import { purposeReadModelServiceBuilder } from "pagopa-interop-readmodel";
+import { upsertTenant } from "pagopa-interop-readmodel/testUtils";
 import { readModelServiceBuilder } from "../src/services/readModelService.js";
 import { readModelServiceBuilderSQL } from "../src/services/readModelServiceSQL.js";
 import { config } from "../src/configs/config.js";
@@ -29,7 +27,6 @@ afterEach(cleanup);
 export const { purposes, tenants } = readModelRepository;
 
 const purposeReadModelServiceSQL = purposeReadModelServiceBuilder(readModelDB);
-const tenantReadModelServiceSQL = tenantReadModelServiceBuilder(readModelDB);
 
 const oldReadModelService = readModelServiceBuilder(readModelRepository);
 const readModelServiceSQL = readModelServiceBuilderSQL(readModelDB);
@@ -51,7 +48,7 @@ export async function seedCollection<T>(
 
 export const seedTenants = async (tenants: Tenant[]): Promise<void> => {
   for (const t of tenants) {
-    await tenantReadModelServiceSQL.upsertTenant(t, 0);
+    await upsertTenant(readModelDB, t, 0);
   }
 };
