+++ packages/readmodel/src/tenantReadModelService.ts
@@ -1,4 +1,4 @@
-import { and, eq, inArray, lte, SQL } from "drizzle-orm";
+import { eq, inArray, SQL } from "drizzle-orm";
 import { Tenant, TenantId, WithMetadata } from "pagopa-interop-models";
 import {
   DrizzleReturnType,
@@ -21,83 +21,11 @@ import {
   TenantVerifiedAttributeVerifierSQL,
 } from "pagopa-interop-readmodel-models";
 import { ascLower } from "pagopa-interop-commons";
-import { splitTenantIntoObjectsSQL } from "./tenant/splitters.js";
 import { aggregateTenant, aggregateTenantArray } from "./tenant/aggregators.js";
-import { checkMetadataVersion } from "./utils.js";
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function tenantReadModelServiceBuilder(db: DrizzleReturnType) {
   return {
-    // eslint-disable-next-line sonarjs/cognitive-complexity
-    async upsertTenant(tenant: Tenant, metadataVersion: number): Promise<void> {
-      const {
-        tenantSQL,
-        mailsSQL,
-        certifiedAttributesSQL,
-        declaredAttributesSQL,
-        verifiedAttributesSQL,
-        verifiedAttributeVerifiersSQL,
-        verifiedAttributeRevokersSQL,
-        featuresSQL,
-      } = splitTenantIntoObjectsSQL(tenant, metadataVersion);
-
-      await db.transaction(async (tx) => {
-        const shouldUpsert = await checkMetadataVersion(
-          tx,
-          tenantInReadmodelTenant,
-          metadataVersion,
-          tenant.id
-        );
-
-        if (!shouldUpsert) {
-          return;
-        }
-
-        await tx
-          .delete(tenantInReadmodelTenant)
-          .where(eq(tenantInReadmodelTenant.id, tenant.id));
-
-        await tx.insert(tenantInReadmodelTenant).values(tenantSQL);
-
-        for (const mailSQL of mailsSQL) {
-          await tx.insert(tenantMailInReadmodelTenant).values(mailSQL);
-        }
-
-        for (const certifiedAttributeSQL of certifiedAttributesSQL) {
-          await tx
-            .insert(tenantCertifiedAttributeInReadmodelTenant)
-            .values(certifiedAttributeSQL);
-        }
-
-        for (const declaredAttributeSQL of declaredAttributesSQL) {
-          await tx
-            .insert(tenantDeclaredAttributeInReadmodelTenant)
-            .values(declaredAttributeSQL);
-        }
-
-        for (const verifiedAttributeSQL of verifiedAttributesSQL) {
-          await tx
-            .insert(tenantVerifiedAttributeInReadmodelTenant)
-            .values(verifiedAttributeSQL);
-        }
-
-        for (const verifierSQL of verifiedAttributeVerifiersSQL) {
-          await tx
-            .insert(tenantVerifiedAttributeVerifierInReadmodelTenant)
-            .values(verifierSQL);
-        }
-
-        for (const revokerSQL of verifiedAttributeRevokersSQL) {
-          await tx
-            .insert(tenantVerifiedAttributeRevokerInReadmodelTenant)
-            .values(revokerSQL);
-        }
-
-        for (const featureSQL of featuresSQL) {
-          await tx.insert(tenantFeatureInReadmodelTenant).values(featureSQL);
-        }
-      });
-    },
     async getTenantById(
       tenantId: TenantId
     ): Promise<WithMetadata<Tenant> | undefined> {
@@ -229,19 +157,8 @@ export function tenantReadModelServiceBuilder(db: DrizzleReturnType) {
         featuresSQL,
       });
     },
-    async deleteTenantById(tenantId: TenantId, version: number): Promise<void> {
-      await db
-        .delete(tenantInReadmodelTenant)
-        .where(
-          and(
-            eq(tenantInReadmodelTenant.id, tenantId),
-            lte(tenantInReadmodelTenant.metadataVersion, version)
-          )
-        );
-    },
   };
 }
-
 export type TenantReadModelService = ReturnType<
   typeof tenantReadModelServiceBuilder
 >;
