+++ packages/readmodel/test/tenantQueries.test.ts
@@ -1,305 +1,102 @@
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
-
 import { describe, it, expect } from "vitest";
-
-import { WithMetadata, Tenant } from "pagopa-interop-models";
-import { sortTenant } from "pagopa-interop-commons-test";
 import {
-  initMockTenant,
-  tenantReadModelService,
-  retrieveTenantSQLObjects,
-} from "./tenantUtils.js";
+  getMockCertifiedTenantAttribute,
+  getMockDeclaredTenantAttribute,
+  getMockTenant,
+  getMockTenantMail,
+  getMockVerifiedTenantAttribute,
+  sortTenant,
+} from "pagopa-interop-commons-test";
+import {
+  DelegationId,
+  generateId,
+  Tenant,
+  tenantFeatureType,
+  tenantKind,
+  TenantRevoker,
+  tenantUnitType,
+  TenantVerifier,
+} from "pagopa-interop-models";
+import { upsertTenant } from "../src/testUtils.js";
+import { tenantReadModelService } from "./tenantUtils.js";
+import { readModelDB } from "./utils.js";
 
 describe("Tenant Queries", () => {
-  describe("Upsert Tenant", () => {
-    it("should add a complete (*all* fields) tenant", async () => {
-      const isTenantComplete = true;
-      const {
-        tenant,
-        tenantForVerifying,
-        tenantForRevoking,
-        tenantMails,
-        tenantCertifiedAttributes,
-        tenantDeclaredAttributes,
-        tenantVerifiedAttributes,
-        tenantFeatures,
-      } = initMockTenant({ isTenantComplete });
-
-      await tenantReadModelService.upsertTenant(
-        tenantForVerifying.data,
-        tenantForVerifying.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenantForRevoking.data,
-        tenantForRevoking.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenant.data,
-        tenant.metadata.version
-      );
-
-      const retrievedTenant = await tenantReadModelService.getTenantById(
-        tenant.data.id
-      );
-
-      const {
-        retrievedTenantSQL,
-        retrievedMailsSQL,
-        retrievedCertifiedAttributesSQL,
-        retrievedDeclaredAttributesSQL,
-        retrievedVerifiedAttributesSQL,
-        retrievedVerifiedAttributeVerifiersSQL,
-        retrievedVerifiedAttributeRevokersSQL,
-        retrievedFeaturesSQL,
-      } = await retrieveTenantSQLObjects(tenant);
-      expect(retrievedTenantSQL).toBeDefined();
-      expect(retrievedMailsSQL).toHaveLength(tenantMails.length);
-      expect(retrievedCertifiedAttributesSQL).toHaveLength(
-        tenantCertifiedAttributes.length
-      );
-      expect(retrievedDeclaredAttributesSQL).toHaveLength(
-        tenantDeclaredAttributes.length
-      );
-      expect(retrievedVerifiedAttributesSQL).toHaveLength(
-        tenantVerifiedAttributes.length
-      );
-      expect(retrievedVerifiedAttributeVerifiersSQL).toHaveLength(1);
-      expect(retrievedVerifiedAttributeRevokersSQL).toHaveLength(1);
-      expect(retrievedFeaturesSQL).toHaveLength(tenantFeatures.length);
-
-      expect(sortTenant(retrievedTenant!)).toStrictEqual(sortTenant(tenant));
-    });
-    it("should add an incomplete (*only* mandatory fields) tenant", async () => {
-      const isTenantComplete = false;
-      const {
-        tenant,
-        tenantForVerifying,
-        tenantForRevoking,
-        tenantMails,
-        tenantCertifiedAttributes,
-        tenantDeclaredAttributes,
-        tenantVerifiedAttributes,
-        tenantFeatures,
-      } = initMockTenant({ isTenantComplete });
-
-      await tenantReadModelService.upsertTenant(
-        tenantForVerifying.data,
-        tenantForVerifying.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenantForRevoking.data,
-        tenantForRevoking.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenant.data,
-        tenant.metadata.version
-      );
-
-      const retrievedTenant = await tenantReadModelService.getTenantById(
-        tenant.data.id
-      );
-
-      const {
-        retrievedTenantSQL,
-        retrievedMailsSQL,
-        retrievedCertifiedAttributesSQL,
-        retrievedDeclaredAttributesSQL,
-        retrievedVerifiedAttributesSQL,
-        retrievedVerifiedAttributeVerifiersSQL,
-        retrievedVerifiedAttributeRevokersSQL,
-        retrievedFeaturesSQL,
-      } = await retrieveTenantSQLObjects(tenant);
-      expect(retrievedTenantSQL).toBeDefined();
-      expect(retrievedMailsSQL).toHaveLength(tenantMails.length);
-      expect(retrievedCertifiedAttributesSQL).toHaveLength(
-        tenantCertifiedAttributes.length
-      );
-      expect(retrievedDeclaredAttributesSQL).toHaveLength(
-        tenantDeclaredAttributes.length
-      );
-      expect(retrievedVerifiedAttributesSQL).toHaveLength(
-        tenantVerifiedAttributes.length
-      );
-      expect(retrievedVerifiedAttributeVerifiersSQL).toHaveLength(1);
-      expect(retrievedVerifiedAttributeRevokersSQL).toHaveLength(1);
-      expect(retrievedFeaturesSQL).toHaveLength(tenantFeatures.length);
-
-      expect(sortTenant(retrievedTenant)).toStrictEqual(sortTenant(tenant));
-    });
-    it("should update a complete (*all* fields) tenant", async () => {
-      const isTenantComplete = true;
-      const {
-        tenant,
-        tenantForVerifying,
-        tenantForRevoking,
-        tenantMails,
-        tenantCertifiedAttributes,
-        tenantDeclaredAttributes,
-        tenantVerifiedAttributes,
-        tenantFeatures,
-      } = initMockTenant({ isTenantComplete });
-      await tenantReadModelService.upsertTenant(
-        tenantForVerifying.data,
-        tenantForVerifying.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenantForRevoking.data,
-        tenantForRevoking.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenant.data,
-        tenant.metadata.version
-      );
-
-      const updatedTenant: WithMetadata<Tenant> = {
-        data: {
-          ...tenant.data,
-          name: "An updated Tenant",
-          updatedAt: new Date(),
-        },
-        metadata: { version: 2 },
-      };
-
-      await tenantReadModelService.upsertTenant(
-        updatedTenant.data,
-        updatedTenant.metadata.version
-      );
-
-      const retrievedTenant = await tenantReadModelService.getTenantById(
-        tenant.data.id
-      );
-
-      const {
-        retrievedTenantSQL,
-        retrievedMailsSQL,
-        retrievedCertifiedAttributesSQL,
-        retrievedDeclaredAttributesSQL,
-        retrievedVerifiedAttributesSQL,
-        retrievedVerifiedAttributeVerifiersSQL,
-        retrievedVerifiedAttributeRevokersSQL,
-        retrievedFeaturesSQL,
-      } = await retrieveTenantSQLObjects(tenant);
-      expect(retrievedTenantSQL).toBeDefined();
-      expect(retrievedMailsSQL).toHaveLength(tenantMails.length);
-      expect(retrievedCertifiedAttributesSQL).toHaveLength(
-        tenantCertifiedAttributes.length
-      );
-      expect(retrievedDeclaredAttributesSQL).toHaveLength(
-        tenantDeclaredAttributes.length
-      );
-      expect(retrievedVerifiedAttributesSQL).toHaveLength(
-        tenantVerifiedAttributes.length
-      );
-      expect(retrievedVerifiedAttributeVerifiersSQL).toHaveLength(1);
-      expect(retrievedVerifiedAttributeRevokersSQL).toHaveLength(1);
-      expect(retrievedFeaturesSQL).toHaveLength(tenantFeatures.length);
-
-      expect(sortTenant(retrievedTenant)).toStrictEqual(
-        sortTenant(updatedTenant)
-      );
-    });
-  });
   describe("Get a Tenant", () => {
     it("should get a tenant from a tenantId", async () => {
-      const isTenantComplete = true;
-      const { tenant, tenantForVerifying, tenantForRevoking } = initMockTenant({
-        isTenantComplete,
-      });
+      const delegationId = generateId<DelegationId>();
+      const tenantVerifier: TenantVerifier = {
+        id: generateId(),
+        verificationDate: new Date(),
+        expirationDate: new Date(),
+        extensionDate: new Date(),
+        delegationId,
+      };
+      const tenantRevoker: TenantRevoker = {
+        id: generateId(),
+        verificationDate: new Date(),
+        revocationDate: new Date(),
+        expirationDate: new Date(),
+        extensionDate: new Date(),
+        delegationId,
+      };
 
-      await tenantReadModelService.upsertTenant(
-        tenantForVerifying.data,
-        tenantForVerifying.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenantForRevoking.data,
-        tenantForRevoking.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenant.data,
-        tenant.metadata.version
-      );
+      const tenant: Tenant = {
+        ...getMockTenant(),
+        kind: tenantKind.PA,
+        updatedAt: new Date(),
+        onboardedAt: new Date(),
+        subUnitType: tenantUnitType.AOO,
+        mails: [getMockTenantMail(), getMockTenantMail()],
+        features: [
+          {
+            type: tenantFeatureType.persistentCertifier,
+            certifierId: generateId(),
+          },
+          {
+            type: tenantFeatureType.delegatedConsumer,
+            availabilityTimestamp: new Date(),
+          },
+          {
+            type: tenantFeatureType.delegatedProducer,
+            availabilityTimestamp: new Date(),
+          },
+        ],
+        attributes: [
+          {
+            ...getMockCertifiedTenantAttribute(),
+            assignmentTimestamp: new Date(),
+            revocationTimestamp: new Date(),
+          },
+          {
+            ...getMockDeclaredTenantAttribute(),
+            assignmentTimestamp: new Date(),
+            revocationTimestamp: new Date(),
+            delegationId,
+          },
+          {
+            ...getMockVerifiedTenantAttribute(),
+            verifiedBy: [tenantVerifier],
+            revokedBy: [tenantRevoker],
+            assignmentTimestamp: new Date(),
+          },
+        ],
+      };
+
+      await upsertTenant(readModelDB, tenant, 1);
 
       const retrievedTenant = await tenantReadModelService.getTenantById(
-        tenant.data.id
+        tenant.id
       );
 
       expect(sortTenant(retrievedTenant)).toStrictEqual(sortTenant(tenant));
     });
     it("should *not* get a tenant from a tenantId", async () => {
-      const { tenant, tenantForVerifying, tenantForRevoking } = initMockTenant({
-        isTenantComplete: true,
-      });
-      await tenantReadModelService.upsertTenant(
-        tenantForVerifying.data,
-        tenantForVerifying.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenantForRevoking.data,
-        tenantForRevoking.metadata.version
-      );
-
       const retrievedTenant = await tenantReadModelService.getTenantById(
-        tenant.data.id
+        generateId()
       );
 
       expect(retrievedTenant).toBeUndefined();
-    });
-  });
-  describe("Delete a Tenant", () => {
-    it("should delete a tenant from a tenantId", async () => {
-      const { tenant, tenantForVerifying, tenantForRevoking } = initMockTenant({
-        isTenantComplete: true,
-      });
-
-      await tenantReadModelService.upsertTenant(
-        tenantForVerifying.data,
-        tenantForVerifying.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenantForRevoking.data,
-        tenantForRevoking.metadata.version
-      );
-      await tenantReadModelService.upsertTenant(
-        tenant.data,
-        tenant.metadata.version
-      );
-
-      const retrievedInsertedTenant =
-        await tenantReadModelService.getTenantById(tenant.data.id);
-
-      await tenantReadModelService.deleteTenantById(
-        tenant.data.id,
-        tenant.metadata.version
-      );
-
-      const deletedInsertedTenant = await tenantReadModelService.getTenantById(
-        tenant.data.id
-      );
-
-      const {
-        retrievedTenantSQL,
-        retrievedMailsSQL,
-        retrievedCertifiedAttributesSQL,
-        retrievedDeclaredAttributesSQL,
-        retrievedVerifiedAttributesSQL,
-        retrievedVerifiedAttributeVerifiersSQL,
-        retrievedVerifiedAttributeRevokersSQL,
-        retrievedFeaturesSQL,
-      } = await retrieveTenantSQLObjects(tenant);
-
-      expect(sortTenant(retrievedInsertedTenant)).toStrictEqual(
-        sortTenant(tenant)
-      );
-
-      expect(deletedInsertedTenant).toBeUndefined();
-      expect(retrievedTenantSQL).toBeUndefined();
-      expect(retrievedMailsSQL).toBeUndefined();
-      expect(retrievedCertifiedAttributesSQL).toBeUndefined();
-      expect(retrievedDeclaredAttributesSQL).toBeUndefined();
-      expect(retrievedVerifiedAttributesSQL).toBeUndefined();
-      expect(retrievedVerifiedAttributeVerifiersSQL).toBeUndefined();
-      expect(retrievedVerifiedAttributeRevokersSQL).toBeUndefined();
-      expect(retrievedFeaturesSQL).toBeUndefined();
     });
   });
 });
