+++ packages/selfcare-client-users-updater/test/utils.ts
@@ -14,10 +14,8 @@ import {
   writeInReadmodel,
 } from "pagopa-interop-commons-test";
 import { afterEach, inject } from "vitest";
-import {
-  clientReadModelServiceBuilder,
-  tenantReadModelServiceBuilder,
-} from "pagopa-interop-readmodel";
+import { clientReadModelServiceBuilder } from "pagopa-interop-readmodel";
+import { upsertTenant } from "pagopa-interop-readmodel/testUtils";
 import { readModelServiceBuilderSQL } from "../src/services/readModelServiceSQL.js";
 import { readModelServiceBuilder } from "../src/services/readModelService.js";
 import { config } from "../src/config/config.js";
@@ -114,5 +112,5 @@ export const addOneClient = async (client: Client): Promise<void> => {
 
 export const addOneTenant = async (tenant: Tenant): Promise<void> => {
   await writeInReadmodel(toReadModelTenant(tenant), tenants);
-  await tenantReadModelServiceBuilder(readModelDB).upsertTenant(tenant, 0);
+  await upsertTenant(readModelDB, tenant, 0);
 };
