+++ packages/tenant-process/test/integrationUtils.ts
@@ -36,6 +36,7 @@ import {
   upsertAgreement,
   upsertAttribute,
   upsertEService,
+  upsertTenant,
 } from "pagopa-interop-readmodel/testUtils";
 import { readModelServiceBuilder } from "../src/services/readModelService.js";
 import { readModelServiceBuilderSQL } from "../src/services/readModelServiceSQL.js";
@@ -133,7 +134,7 @@ export const addOneAttribute = async (attribute: Attribute): Promise<void> => {
 export const addOneTenant = async (tenant: Tenant): Promise<void> => {
   await writeTenantInEventstore(tenant);
   await writeInReadmodel(toReadModelTenant(tenant), tenants);
-  await tenantReadModelServiceSQL.upsertTenant(tenant, 0);
+  await upsertTenant(readModelDB, tenant, 0);
 };
 
 export const addOneDelegation = async (
