+++ packages/tenant-readmodel-writer-sql/src/consumerServiceV1.ts
@@ -5,11 +5,11 @@ import {
   missingKafkaMessageDataError,
   unsafeBrandId,
 } from "pagopa-interop-models";
-import { ReadModelService } from "./readModelService.js";
+import { TenantWriterService } from "./tenantWriterService.js";
 
 export async function handleMessageV1(
   message: TenantEventEnvelopeV1,
-  readModelService: ReadModelService
+  tenantWriterService: TenantWriterService
 ): Promise<void> {
   await match(message)
     .with(
@@ -21,20 +21,20 @@ export async function handleMessageV1(
           throw missingKafkaMessageDataError("tenant", msg.type);
         }
 
-        await readModelService.upsertTenant(
+        await tenantWriterService.upsertTenant(
           fromTenantV1(msg.data.tenant),
           message.version
         );
       }
     )
     .with({ type: "TenantDeleted" }, async (msg) => {
-      await readModelService.deleteTenant(
+      await tenantWriterService.deleteTenantById(
         unsafeBrandId(msg.data.tenantId),
         message.version
       );
     })
     .with({ type: "SelfcareMappingCreated" }, async (msg) => {
-      await readModelService.setSelfcareId(
+      await tenantWriterService.setSelfcareId(
         unsafeBrandId(msg.data.tenantId),
         msg.data.selfcareId,
         msg.version
@@ -42,7 +42,7 @@ export async function handleMessageV1(
     })
     .with({ type: "SelfcareMappingDeleted" }, async () => Promise.resolve())
     .with({ type: "TenantMailDeleted" }, async (msg) => {
-      await readModelService.deleteTenantMailById(
+      await tenantWriterService.deleteTenantMailById(
         unsafeBrandId(msg.data.tenantId),
         msg.data.mailId,
         msg.version
